requirements.txt
streamlit
Pillow
docx
pikepdf
